simple-post-server/requirements.txt
influxdb-client
paho-mqtt
pyYAML
requests
